Retire Netlify: relocate shared AutoBot modules to shared/, delete netlify/ + netlify.toml (#864)
Option B (full retirement) per owner's issue decision. netlify/functions/
was not dead code: functions/api/gemini-proxy.js (the live Cloudflare
AutoBot proxy) and the build-time index generators (scripts/build-autobot-
index.mjs, scripts/build-llms-txt.mjs) both import from it.

- Relocate the 5 live modules (autobot-core.mjs, github-context.mjs,
  docs-retrieval.mjs, docs-loader.mjs, constants.mjs) to a new top-level
  shared/ directory -- not functions/, which is a routing-sensitive path
  for this repo's Cloudflare Pages Functions convention. shared/ reflects
  that these modules serve both the worker runtime and build-time scripts,
  not just one.
- Delete the dead netlify/functions/gemini-proxy.mjs handler (superseded by
  functions/api/gemini-proxy.js; nothing imported it).
- Update every importer: functions/api/gemini-proxy.js, scripts/build-
  autobot-index.mjs, scripts/build-llms-txt.mjs, tests/docs-loader.test.js,
  tests/enhanced-instruction.test.js, demo-autobot-approach.mjs.
- Fix shared/docs-loader.mjs's currentDir-relative paths to snippets.json
  and docs/ for the new (one level shallower) location.
- Delete the now-empty netlify/ directory and netlify.toml (its redirects
  were already fully mirrored by docusaurus.config.js's
  plugin-client-redirects).
- Sync docs/maintainers/site-operations.md and README.md to drop retired
  Netlify build/hosting references; update docusaurus.config.js's redirect
  comment to stop citing netlify.toml as authoritative.

Verified: yarn build (regenerates static/autobot-index.json, 1,992,757
bytes), yarn test, yarn typecheck, yarn test:playwright (12/12), docs-
quality.test.js, and check-doc-duplicates.mjs all green.

Closes #864
