--- netlify.toml
@@ -1,203 +1,0 @@
-[build]
-  command = "yarn build"
-  publish = "build"
-
-[build.environment]
-  NODE_VERSION = "20"
-
-[[redirects]]
-  from = "/api/*"
-  to = "/.netlify/functions/:splat"
-  status = 200
-
-[[redirects]]
-  from = "/docs/Keywords/*"
-  to = "/docs/reference/actions/:splat"
-  status = 301
-
-[[redirects]]
-  from = "/docs/Basic_Config/*"
-  to = "/docs/reference/configuration/:splat"
-  status = 301
-
-[[redirects]]
-  from = "/docs/reference/configuration/basicConfig"
-  to = "/docs/reference/configuration/webConfig"
-  status = 301
-
-[[redirects]]
-  from = "/docs/reference/configuration/basicConfig2"
-  to = "/docs/reference/configuration/mobileConfig"
-  status = 301
-
-[[redirects]]
-  from = "/docs/reference/configuration/basicConfig3"
-  to = "/docs/reference/configuration/apiConfig"
-  status = 301
-
-[[redirects]]
-  from = "/docs/Best_Practices/*"
-  to = "/docs/reference/guides/:splat"
-  status = 301
-
-[[redirects]]
-  from = "/docs/Properties/*"
-  to = "/docs/reference/properties/:splat"
-  status = 301
-
-[[redirects]]
-  from = "/docs/Reporting/*"
-  to = "/docs/reference/reporting/:splat"
-  status = 301
-
-[[redirects]]
-  from = "/docs/Getting_Started/first_steps"
-  to = "/docs/start/overview"
-  status = 301
-
-[[redirects]]
-  from = "/docs/Getting_Started/first_steps_2"
-  to = "/docs/start/quick-start"
-  status = 301
-
-[[redirects]]
-  from = "/docs/Getting_Started/first_steps_3"
-  to = "/docs/start/quick-start"
-  status = 301
-
-[[redirects]]
-  from = "/docs/Getting_Started/first_steps_4"
-  to = "/docs/start/installation"
-  status = 301
-
-[[redirects]]
-  from = "/docs/Getting_Started/first_steps_5"
-  to = "/docs/start/installation"
-  status = 301
-
-[[redirects]]
-  from = "/docs/Getting_Started/first_steps_6"
-  to = "/docs/start/installation"
-  status = 301
-
-[[redirects]]
-  from = "/docs/Getting_Started/shaft_wizard"
-  to = "/docs/start/installation"
-  status = 301
-
-[[redirects]]
-  from = "/docs/Getting_Started/setup_web"
-  to = "/docs/testing/web"
-  status = 301
-
-[[redirects]]
-  from = "/docs/Getting_Started/setup_mobile"
-  to = "/docs/testing/mobile"
-  status = 301
-
-[[redirects]]
-  from = "/docs/Getting_Started/flutter_testing"
-  to = "/docs/testing/mobile"
-  status = 301
-
-[[redirects]]
-  from = "/docs/Getting_Started/setup_api"
-  to = "/docs/testing/api"
-  status = 301
-
-[[redirects]]
-  from = "/docs/Getting_Started/integrations"
-  to = "/docs/features/modules"
-  status = 301
-
-[[redirects]]
-  from = "/docs/Getting_Started/support"
-  to = "/docs/start/overview"
-  status = 301
-
-[[redirects]]
-  from = "/docs/Getting_Started/JUnit5_Integration"
-  to = "/docs/reference/guides/JUnit_Integration"
-  status = 301
-
-[[redirects]]
-  from = "/docs/Demos/web"
-  to = "/docs/testing/web"
-  status = 301
-
-[[redirects]]
-  from = "/docs/Demos/mobile"
-  to = "/docs/testing/mobile"
-  status = 301
-
-# Release blog posts published before 2026-06-19 were consolidated into a
-# single chronological history post (issue #841). Redirect the retired
-# standalone slugs to the section covering the same release.
-[[redirects]]
-  from = "/blog/release_announcement_7.1.20230309"
-  to = "/blog/release-history"
-  status = 301
-
-[[redirects]]
-  from = "/blog/release-10.1.20260324"
-  to = "/blog/release-history"
-  status = 301
-
-[[redirects]]
-  from = "/blog/release-10.1.20260331"
-  to = "/blog/release-history"
-  status = 301
-
-[[redirects]]
-  from = "/blog/release-10.2.20260501"
-  to = "/blog/release-history"
-  status = 301
-
-[[redirects]]
-  from = "/blog/release-10.2.20260505"
-  to = "/blog/release-history"
-  status = 301
-
-[[redirects]]
-  from = "/blog/release-10.2.20260506"
-  to = "/blog/release-history"
-  status = 301
-
-[[redirects]]
-  from = "/blog/release-10.2.20260605"
-  to = "/blog/release-history"
-  status = 301
-
-[[redirects]]
-  from = "/blog/release-10.2.20260610"
-  to = "/blog/release-history"
-  status = 301
-
-[[redirects]]
-  from = "/blog/release-10.2.20260612"
-  to = "/blog/release-history"
-  status = 301
-
-[[redirects]]
-  from = "/blog/release-10.2.20260614"
-  to = "/blog/release-history"
-  status = 301
-
-[[redirects]]
-  from = "/blog/release-10.2.20260615"
-  to = "/blog/release-history"
-  status = 301
-
-[[redirects]]
-  from = "/blog/release-10.2.20260617"
-  to = "/blog/release-history"
-  status = 301
-
-[[redirects]]
-  from = "/blog/release-10.2.20260618"
-  to = "/blog/release-history"
-  status = 301
-
-[functions]
-  directory = "netlify/functions"
-  included_files = ["docs/**", "src/data/snippets.json"]
