+++ scripts/build-llms-txt.mjs
@@ -1,7 +1,7 @@
 import {mkdir, readFile, writeFile} from 'node:fs/promises';
 import path from 'node:path';
 import {fileURLToPath} from 'node:url';
-import {buildDocumentationIndex} from '../netlify/functions/docs-loader.mjs';
+import {buildDocumentationIndex} from '../shared/docs-loader.mjs';
 
 const root = path.resolve(path.dirname(fileURLToPath(import.meta.url)), '..');
 const docsRoot = path.join(root, 'docs');
