+++ shared/autobot-core.mjs
